rename foundations to tokens
