+++ apps/storybook/stories/design/design-tokens/Borders.tsx
@@ -1,5 +1,5 @@
 import { Table, Thead, Tbody, Tr, Th, Td, TableContainer, Code } from "@chakra-ui/react";
-import { borders } from "@kvib/react/src/theme/foundations";
+import { borders } from "@kvib/react/src/theme/tokens";
 
 export const Borders = () => {
   return (
