+++ apps/storybook/stories/design/design-tokens/Breakpoints.tsx
@@ -1,5 +1,5 @@
 import { Table, Thead, Tbody, Tr, Th, Td, TableContainer, Code } from "@chakra-ui/react";
-import { breakpoints } from "@kvib/react/src/theme/foundations";
+import { breakpoints } from "@kvib/react/src/theme/tokens";
 
 export const Breakpoints = () => {
   return (
