+++ packages/react/src/theme/index.ts
@@ -1,11 +1,11 @@
 import { extendTheme } from "@chakra-ui/react";
 import { theme as defaultTheme } from "@chakra-ui/theme";
 import * as components from "./components";
-import * as foundations from "./foundations";
+import * as tokens from "./tokens";
 
 export const theme = extendTheme({
   ...defaultTheme,
-  ...foundations,
+  ...tokens,
   components: {
     ...defaultTheme.components,
     ...components,
